Fixing some data models and removing circular dependency between services and ui packages
