+++ libs/designer-ui/src/lib/monitoring/values/types.ts
@@ -1,21 +1,3 @@
-export interface ContentHash {
-  algorithm: string;
-  value: string;
-}
-
-export interface ContentLink {
-  contentHash: ContentHash;
-  contentSize: number;
-  contentVersion: string;
-  metadata?: Record<string, unknown>;
-  secureData?: SecureData;
-  uri?: string;
-}
-
-export interface SecureData {
-  properties: string[];
-}
-
 export interface ValueProps {
   displayName: string;
   format?: string;
