--- libs/designer-ui/src/lib/monitoring/valuespanel/types.ts
@@ -1,10 +1,0 @@
-export type BoundParameters = Record<string, BoundParameter<any>>;
-
-interface BoundParameter<T> {
-  displayName: string;
-  dynamicValue?: T;
-  format?: string;
-  language?: string;
-  value: T;
-  visibility?: string;
-}
