+++ libs/designer-ui/src/lib/overview/overviewcommandbar.tsx
@@ -1,7 +1,7 @@
-import type { CallbackInfo } from './types';
 import { getCallbackUrl } from './utils';
 import type { ICommandBarItemProps } from '@fluentui/react';
 import { CommandBar } from '@fluentui/react';
+import type { CallbackInfo } from '@microsoft/utils-logic-apps';
 import { useMemo } from 'react';
 import { useIntl } from 'react-intl';
