+++ libs/designer-ui/src/lib/overview/utils.ts
@@ -1,7 +1,7 @@
 import { getDurationString } from '../utils';
-import type { CallbackInfo, Run, RunDisplayItem, RunError } from './types';
-import { isCallbackInfoWithRelativePath } from './types';
-import { isObject } from '@microsoft/utils-logic-apps';
+import type { RunDisplayItem } from './types';
+import type { CallbackInfo, Run, RunError } from '@microsoft/utils-logic-apps';
+import { isCallbackInfoWithRelativePath, isObject } from '@microsoft/utils-logic-apps';
 
 export function getCallbackUrl(callbackInfo: CallbackInfo | undefined): string | undefined {
   if (!callbackInfo) {
