+++ libs/services/designer-client-services/src/lib/run.ts
@@ -1,6 +1,5 @@
 import type { CallbackInfo } from './workflow';
-import type { ContentLink, Runs, RunError, Run } from '@microsoft/designer-ui';
-import type { LogicAppsV2 } from '@microsoft/utils-logic-apps';
+import type { ContentLink, LogicAppsV2, Run, RunError, Runs } from '@microsoft/utils-logic-apps';
 import { AssertionErrorCode, AssertionException } from '@microsoft/utils-logic-apps';
 
 export interface IRunService {
