+++ libs/utils/src/lib/models/index.ts
@@ -7,6 +7,7 @@ export * from './operationDiscoveryResults';
 export * from './operationmanifest';
 export * from './workflowNode';
 export * from './gateway';
+export * from './run';
 export * from './subscription';
 export * from './theme';
